correct backend with admin

+++ routes/rental_route.js
@@ -1,15 +1,22 @@
 const express = require("express");
-const { findAll, save, findUserBookings, findById, deleteById, update } = require("../controller/rental_controller");
+const { findAll, save, findUserBookings, findById, deleteById, update, getAdminBookings, updateBooking, deleteBooking, processPayment } = require("../controller/rental_controller");
 
 const router = express.Router();
 
 router.get("/findAll", findAll)
+router.get("/userBookings", findUserBookings);
+router.get("/adminBookings", getAdminBookings);
+router.put("/updateBooking/:id", updateBooking);
+router.delete("/deleteBooking/:id", deleteBooking);
+
+router.post("/processPayment", processPayment);
+
 router.post("/", save)
 router.get("/:id", findById)
 router.delete("/:id", deleteById)
 router.put("/:id", update)
 
-router.get("/userBookings", findUserBookings); // No need to use a parameter
+
